Cleaned up backend advice. Moved to LocalDate updated date formatting in the frontend added new controller methods

+++ src/main/java/edu/team5/wright_time/model/repository/AircraftRepository.java
@@ -3,6 +3,6 @@ package edu.team5.wright_time.model.repository;
 import edu.team5.wright_time.model.entity.Aircraft;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AircraftRepository extends CrudRepository<Aircraft,Integer> {
+public interface AircraftRepository extends CrudRepository<Aircraft, Long> {
     
 }
